Lab05/5.4/poetry-demo/requirements.txt
poetry==1.1.4
kafka-python@latest